Refactored the mutation generator as a static function. Reworked the way mutations are loaded as a first step to have a pluggable API
